+++ fastserver/modules/sys/dao/menu_dao.go
@@ -9,7 +9,7 @@ type MenuDao struct {
 }
 
 // 批量删除菜单
-func (m *MenuDao) BatchDeleteMenuByIds(menuIds []uint) error {
+func (m *MenuDao) BatchDeleteMenuByIds(menuIds []uint64) error {
 	menus, err := database.GetByIds[model.Menu](menuIds)
 	err = database.DB.Select("Roles").Unscoped().Delete(&menus).Error
 	return err
